requirements.txt
agno
aiohttp
certifi
duckduckgo_search
httpx
lancedb
nest_asyncio
numpy
openai
pandas
Pillow
primp
pyarrow
pydantic
python-dotenv
Requests
rich
streamlit
urllib3
mcp
langtrace-python-sdk
deprecated